feat(components/atom/switch): a11y and icons
