+++ components/atom/switch/demo/settings.js
@@ -1,0 +1,10 @@
+export const BASE_CLASS_DEMO = `DemoAtomSwitch`
+export const CLASS_SECTION = `${BASE_CLASS_DEMO}-section`
+
+export const flexCenteredStyle = {
+  display: 'flex',
+  justifyContent: 'center',
+  wrap: 'nowrap',
+  alignItems: 'center',
+  alignContent: 'center'
+}
