+++ components/atom/switch/src/SwitchType/helpers.js
@@ -6,7 +6,7 @@ const CLASS_FULL_WIDTH = `${BASE_CLASS}--fullWidth`
 
 export function switchClassNames({size, classType, fullWidth}) {
   const CLASS_TYPE = `${BASE_CLASS}-${classType}`
-  const CLASS_SIZE = `${BASE_CLASS}--${size}`
+  const CLASS_SIZE = `${BASE_CLASS}--size-${size}`
 
   const className = cx(BASE_CLASS, CLASS_TYPE, CLASS_SIZE, {
     [CLASS_FULL_WIDTH]: fullWidth
