+++ components/atom/switch/src/styles/settings.scss
@@ -1,22 +1,34 @@
+/* SMALL SIZE */
+$h-atom-switch-small: $sz-base * 2 !default;
+$h-atom-switch-container-small: $h-atom-switch-small + 10px !default;
+$w-atom-switch-small: $sz-base * 3 !default;
+$w-atom-switch-circle-small: $h-atom-switch-small !default;
+$m-atom-switch-circle-small--active: $w-atom-switch-small -
+  $w-atom-switch-circle-small !default;
+$h-atom-switch-input-container-small: $h-atom-switch-small !default;
+
 /* DEFAULT SIZE */
-$h-atom-switch-container-default: 34px !default;
-$h-atom-switch-default: 24px !default;
-$w-atom-switch-default: 40px !default;
-$w-atom-switch-circle-default: 22px !default;
-$m-atom-switch-circle-default--active: $m-l !default;
+$h-atom-switch-default: $sz-base * 3 !default;
+$h-atom-switch-container-default: $h-atom-switch-default + 10px !default;
+$w-atom-switch-default: $sz-base * 5 !default;
+$w-atom-switch-circle-default: $h-atom-switch-default !default;
+$m-atom-switch-circle-default--active: $w-atom-switch-default -
+  $w-atom-switch-circle-default !default;
 $p-atom-switch-container: $p-s !default;
 
 /* LARGE SIZE */
-$h-atom-switch-container-large: 42px !default;
-$h-atom-switch-large: 32px !default;
-$w-atom-switch-large: 56px !default;
-$w-atom-switch-circle-large: 32px !default;
-$m-atom-switch-circle-large--active: $m-xl !default;
-$h-atom-switch-input-container-large: 32px !default;
+$h-atom-switch-large: $sz-base * 4 !default;
+$h-atom-switch-container-large: $h-atom-switch-large + 10px !default;
+$w-atom-switch-large: $sz-base * 7 !default;
+$w-atom-switch-circle-large: $h-atom-switch-large !default;
+$m-atom-switch-circle-large--active: $w-atom-switch-large -
+  $w-atom-switch-circle-large !default;
+$h-atom-switch-input-container-large: $h-atom-switch-large !default;
 
 $c-atom-switch--active: color-variation($c-primary, 1) !default;
 $c-atom-switch--active-shadow: rgba($c-black, 0.2);
 $c-atom-switch--default: color-variation($c-gray, 3) !default;
+$c-atom-switch-circle: $c-white;
 
 $bd-atom-switch-default: $bdw-s solid
   color-variation($c-atom-switch--default, $c-dark-step) !default;
